refactor: extract profile fetching logic into custom hook for better testability
Extract complex profile fetching logic from the profile component into a reusable
custom hook to improve separation of concerns, testability, and maintainability.

## New Custom Hook

### useProfileFetching Hook (`src/app/hooks/useProfileFetching.ts`)
- Encapsulates all profile fetching logic and state management
- Handles both own profile and other users' profiles seamlessly
- Implements debounced fetching for other users (300ms configurable delay)
- Immediate fetching for own profile (no debouncing needed)
- Race condition prevention using fetch ID tracking
- Comprehensive error handling with specific error types
- Automatic cleanup on component unmount
- Provides utility functions: `refetchProfile()` and `clearError()`

### Interface Design
```typescript
interface UseProfileFetchingReturn {
  viewedProfile: UserProfile | null
  stats: UserStats | null
  loading: boolean
  error: string | null
  isOwnProfile: boolean
  refetchProfile: () => void
  clearError: () => void
}
```

### Key Features
- **Smart Fetching Strategy**: Immediate for own profile, debounced for others
- **Race Condition Safety**: Prevents stale data from overwriting newer fetches
- **Memory Leak Prevention**: Proper cleanup of pending async operations
- **Error Boundary Integration**: Handles InvalidUIDError and network failures
- **Performance Optimized**: Memoized callbacks and debounced external calls

## Component Refactoring

### Profile Component Updates (`src/app/profile/[[...uid]]/page.tsx`)
- Simplified component logic by 200+ lines
- Removed complex useEffect chains and race condition handling
- Cleaner separation between URL params handling and profile fetching
- Maintained all existing functionality and user experience
- Better error state management with combined params and fetch errors

### Architecture Benefits
- **Testability**: Hook can be tested in isolation from React components
- **Reusability**: Can be used in other components requiring profile data
- **Maintainability**: Single source of truth for profile fetching logic
- **Debugging**: Easier to debug isolated hook vs complex component logic

## Testing & Quality

### Basic Test Coverage
- Hook interface validation
- Function signature verification
- Linting compliance maintained
- Build verification successful

### Documentation Updates
- Updated CLAUDE.md with new hook architecture
- Documented profile routing improvements
- Added custom hook to development guide

This refactoring improves code organization while maintaining all existing
functionality including security validation, performance optimizations,
and error handling.
