+++ src/app/profile/[[...uid]]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { useEffect, useState, useCallback, useMemo, useRef } from 'react'
+import React, { useEffect, useState } from 'react'
 import {
   Camera,
   Edit3,
@@ -17,19 +17,11 @@ import {
   Sparkles,
   Gamepad2,
 } from 'lucide-react'
-import { UserStats, UserProfile } from '@/lib/definitions/user'
-import {
-  getInitials,
-  getMediaIcon,
-  debounce,
-  isValidUID,
-  InvalidUIDError,
-} from '@/lib/utils'
+import { getInitials, getMediaIcon, isValidUID } from '@/lib/utils'
 import { useUserStore } from '@/app/store/user-store'
-import { UserService } from '@/app/services/user-service'
 import { db } from '@/lib/firebase'
 import { doc, setDoc, serverTimestamp } from 'firebase/firestore'
-import { useRankings } from '@/app/hooks/useRankings'
+import { useProfileFetching } from '@/app/hooks/useProfileFetching'
 import Image from 'next/image'
 import { ProfilePageSkeleton } from '@/app/ui/skeletons'
 import Link from 'next/link'
@@ -41,10 +33,10 @@ interface ProfilePageProps {
 
 const ProfilePage = ({ params }: ProfilePageProps): React.ReactElement => {
   // Parse the uid from params - if no uid provided, show current user's profile
-  const [targetUserId, setTargetUserId] = React.useState<string | undefined>(
-    undefined
-  )
-  const [paramsLoaded, setParamsLoaded] = React.useState(false)
+  const [targetUserId, setTargetUserId] = useState<string | undefined>(undefined)
+  const [paramsLoaded, setParamsLoaded] = useState(false)
+  const [paramsError, setParamsError] = useState<string | null>(null)
+  
   const currentUserProfile = useUserStore((state) => state.userProfile)
   const updateUserProfile = useUserStore((state) => state.updateUserProfile)
 
@@ -59,7 +51,7 @@ const ProfilePage = ({ params }: ProfilePageProps): React.ReactElement => {
           // Validate the UID for security
           if (!isValidUID(rawUid)) {
             console.warn('Invalid UID detected:', rawUid)
-            setError('Invalid user ID format')
+            setParamsError('Invalid user ID format')
             setParamsLoaded(true)
             return
           }
@@ -71,173 +63,39 @@ const ProfilePage = ({ params }: ProfilePageProps): React.ReactElement => {
         setParamsLoaded(true)
       } catch (err) {
         console.error('Error loading params:', err)
-        setError('Failed to load profile parameters')
+        setParamsError('Failed to load profile parameters')
         setParamsLoaded(true)
       }
     }
     loadParams()
   }, [params])
 
-  // Determine if we're viewing our own profile or someone else's
-  const isOwnProfile = !targetUserId || targetUserId === currentUserProfile?.uid
+  // Use the profile fetching hook
+  const {
+    viewedProfile,
+    stats,
+    loading,
+    error: fetchError,
+    isOwnProfile,
+  } = useProfileFetching({
+    targetUserId,
+    paramsLoaded,
+    debounceMs: 300,
+  })
 
-  // State for the profile being viewed (could be current user or other user)
-  const [viewedProfile, setViewedProfile] = useState<UserProfile | null>(
-    isOwnProfile ? currentUserProfile : null
-  )
+  // Combine params error and fetch error
+  const error = paramsError || fetchError
+
+  // Bio editing state (only used for own profile)
   const [isEditingBio, setIsEditingBio] = useState(false)
   const [tempBio, setTempBio] = useState('')
-  const [stats, setStats] = useState<UserStats | null>(null)
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
 
-  // Use rankings hook - will need to be enhanced to support other users in Phase 3
-  const { getUserStats, rankings } = useRankings()
-
-  // Keep track of the current fetch to avoid race conditions
-  const currentFetchRef = useRef<string | null>(null)
-
-  // Memoized profile fetching function to prevent unnecessary recreations
-  const fetchProfileData = useCallback(
-    async (userId?: string, isOwn: boolean = false) => {
-      // Don't fetch data until params are loaded
-      if (!paramsLoaded) {
-        return
-      }
-
-      // Create a unique identifier for this fetch
-      const fetchId = `${userId || 'own'}-${Date.now()}`
-      currentFetchRef.current = fetchId
-
-      try {
-        setLoading(true)
-        setError(null)
-
-        let profileToView: UserProfile | null = null
-
-        if (isOwn) {
-          // Use current user's profile from store
-          profileToView = currentUserProfile
-        } else {
-          // Fetch other user's profile
-          if (userId) {
-            try {
-              profileToView = await UserService.getUserProfile(userId)
-
-              // Check if this fetch is still the current one (prevent race conditions)
-              if (currentFetchRef.current !== fetchId) {
-                return // Abort if a newer fetch has started
-              }
-
-              if (!profileToView) {
-                setError('User not found')
-                setLoading(false)
-                return
-              }
-            } catch (fetchError) {
-              // Check if this fetch is still the current one
-              if (currentFetchRef.current !== fetchId) {
-                return
-              }
-
-              if (fetchError instanceof InvalidUIDError) {
-                setError('Invalid user ID format')
-              } else {
-                setError('Failed to load user profile')
-              }
-              setLoading(false)
-              return
-            }
-          }
-        }
-
-        // Final race condition check before updating state
-        if (currentFetchRef.current !== fetchId) {
-          return
-        }
-
-        if (profileToView) {
-          setViewedProfile(profileToView)
-          setTempBio(profileToView.bio || '')
-
-          // For now, only get stats for current user - will enhance in Phase 3
-          if (isOwn) {
-            const userStats = await getUserStats()
-
-            // Final check before setting stats
-            if (currentFetchRef.current === fetchId) {
-              setStats(userStats)
-            }
-          } else {
-            // Placeholder for other users' stats - will implement in Phase 3
-            setStats(null)
-          }
-        }
-      } catch (err) {
-        // Only update error if this is still the current fetch
-        if (currentFetchRef.current === fetchId) {
-          console.error('Error fetching profile data:', err)
-          setError('Failed to load profile')
-        }
-      } finally {
-        // Only update loading if this is still the current fetch
-        if (currentFetchRef.current === fetchId) {
-          setLoading(false)
-        }
-      }
-    },
-    [paramsLoaded, currentUserProfile, getUserStats]
-  )
-
-  // Create debounced version of fetchProfileData
-  const debouncedFetchProfile = useMemo(
-    () => debounce(fetchProfileData, 300),
-    [fetchProfileData]
-  )
-
-  // Effect to trigger profile fetching when dependencies change
+  // Update tempBio when viewedProfile changes
   useEffect(() => {
-    if (!paramsLoaded) {
-      return
+    if (viewedProfile) {
+      setTempBio(viewedProfile.bio || '')
     }
-
-    // For own profile, fetch immediately (no debouncing needed)
-    if (isOwnProfile) {
-      fetchProfileData(undefined, true)
-    } else {
-      // For other users' profiles, use debouncing to prevent excessive API calls
-      if (targetUserId) {
-        debouncedFetchProfile(targetUserId, false)
-      }
-    }
-  }, [
-    paramsLoaded,
-    targetUserId,
-    isOwnProfile,
-    fetchProfileData,
-    debouncedFetchProfile,
-  ])
-
-  // Also trigger when current user profile or rankings change (for own profile)
-  useEffect(() => {
-    if (paramsLoaded && isOwnProfile) {
-      fetchProfileData(undefined, true)
-    }
-  }, [
-    currentUserProfile,
-    rankings,
-    paramsLoaded,
-    isOwnProfile,
-    fetchProfileData,
-  ])
-
-  // Cleanup effect to cancel pending debounced calls on unmount
-  useEffect(() => {
-    return () => {
-      // Clear any pending fetch to prevent state updates after unmount
-      currentFetchRef.current = null
-    }
-  }, [])
+  }, [viewedProfile])
 
   const handleSaveBio = async () => {
     if (!currentUserProfile || !isOwnProfile) return
@@ -252,7 +110,6 @@ const ProfilePage = ({ params }: ProfilePageProps): React.ReactElement => {
       { merge: true }
     )
     updateUserProfile({ ...currentUserProfile, bio: tempBio })
-    setViewedProfile({ ...currentUserProfile, bio: tempBio })
 
     setIsEditingBio(false)
   }
